+++ src/components/Checkbox/Checkbox.stories.ts
@@ -2,25 +2,53 @@ import type { Meta, StoryObj } from "@storybook/react";
 
 import { Checkbox } from "./index";
 
-// More on how to set up stories at: https://storybook.js.org/docs/react/writing-stories/introduction#default-export
 const meta = {
   title: "Components/Checkbox",
   component: Checkbox,
   parameters: {
-    // Optional parameter to center the component in the Canvas. More info: https://storybook.js.org/docs/react/configure/story-layout
     layout: "centered",
+    docs: {
+      subtitle: "A controlled checkbox component.",
+    },
   },
-  // This component will have an automatically generated Autodocs entry: https://storybook.js.org/docs/react/writing-docs/autodocs
   tags: ["autodocs"],
 } satisfies Meta<typeof Checkbox>;
 
 export default meta;
 type Story = StoryObj<typeof meta>;
 
-// More on writing stories with args: https://storybook.js.org/docs/react/writing-stories/args
-export const Primary: Story = {
+export const Default: Story = {
   args: {
     id: "checkbox-1",
+    value: "checkbox1",
     label: "Checkbox",
+  },
+};
+
+export const Checked: Story = {
+  args: {
+    ...Default.args,
+    checked: true,
+  },
+};
+
+export const Unchecked: Story = {
+  args: {
+    ...Default.args,
+  },
+};
+
+export const Indeterminate: Story = {
+  args: {
+    ...Default.args,
+    checked: true,
+    indeterminate: true,
+  },
+};
+
+export const Disabled: Story = {
+  args: {
+    ...Default.args,
+    disabled: true,
   },
 };