+++ src/components/Checkbox/types.ts
@@ -3,10 +3,10 @@ import { ChangeEvent } from "react";
 export type CheckboxProps = {
   id: string;
   value: string;
-  onChange: (event: ChangeEvent<HTMLInputElement>) => void;
+  checked?: boolean;
+  onChange?: (event: ChangeEvent<HTMLInputElement>) => void;
   name?: string;
   label?: string;
-  checked?: boolean;
   disabled?: boolean;
   indeterminate?: boolean;
   classNames?: string;