+++ src/components/TreeView/storybook/TreeView.stories.ts
@@ -9,13 +9,12 @@ const meta = {
   parameters: {
     layout: "centered",
   },
-  tags: ["autodocs"],
 } satisfies Meta<typeof TreeView>;
 
 export default meta;
 type Story = StoryObj<typeof meta>;
 
-export const Primary: Story = {
+export const Default: Story = {
   args: {
     data: STORYBOOK_TREE_VIEW_DATA,
   },